platform: i2c: update header file
Rename `i2c_extra` to platform specific header file.

Update projects accordingly.
