+++ projects/adv7511/src.mk
@@ -50,7 +50,7 @@ INCS += $(DRIVERS)/axi_core/axi_dmac/axi_dmac.h \
 	$(DRIVERS)/axi_core/clk_axi_clkgen/clk_axi_clkgen.h
 INCS +=	$(PLATFORM_DRIVERS)/spi_extra.h \
 	$(PLATFORM_DRIVERS)/gpio_extra.h \
-	$(PLATFORM_DRIVERS)/i2c_extra.h \
+	$(PLATFORM_DRIVERS)/xilinx_i2c.h \
 	$(PLATFORM_DRIVERS)/irq_extra.h \
 	$(PLATFORM_DRIVERS)/timer_extra.h
 INCS +=	$(INCLUDE)/no_os_axi_io.h \
