+++ projects/eval-adxl367z/src/platform/xilinx/platform_src.mk
@@ -3,7 +3,7 @@ SRCS += $(PLATFORM_DRIVERS)/xilinx_spi.c \
 	$(PLATFORM_DRIVERS)/xilinx_delay.c
 
 INCS +=	$(PLATFORM_DRIVERS)/spi_extra.h \
-	$(PLATFORM_DRIVERS)/i2c_extra.h
+	$(PLATFORM_DRIVERS)/xilinx_i2c.h
 
 ifeq (y,$(strip $(IIO_EXAMPLE)))
